* all: Eliminating warnings (mainly from Win32 builds).

+++ exult/trunk/tools/expack.cc
@@ -359,8 +359,8 @@ int main(int argc, char **argv)
 	case EXTRACT: {
 		if (argc == 4) {
 			U7object f(fname, atoi(argv[3]));
-			size_t nobjs = f.number_of_objects();
-			size_t n = strtoul(argv[3], 0, 0);
+			unsigned long nobjs = f.number_of_objects();
+			unsigned long n = strtoul(argv[3], 0, 0);
 			if (n >= nobjs) {
 				cerr << "Obj. #(" << n <<
 				     ") is too large.  ";
